split prefetcher into single file
